moves call_loopy to loopy.py; removes hard dependency on loopy
