+++ pytato/__init__.py
@@ -42,8 +42,6 @@ from pytato.array import (
         equal, not_equal, less, less_equal, greater, greater_equal,
 
         logical_or, logical_and, logical_not,
-
-        call_loopy,
         )
 
 from pytato.target.loopy.codegen import generate_loopy
@@ -59,8 +57,6 @@ __all__ = (
         "make_data_wrapper",
 
         "matmul", "roll", "transpose", "stack", "reshape", "concatenate",
-
-        "call_loopy",
 
         "generate_loopy",
